# api/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Feral File Indexer API
  description: |
    REST API for the Feral File blockchain indexer service.
    
    This API provides access to indexed NFT token data from multiple blockchains,
    including ownership information, metadata, and provenance events.
    
    ## Supported Blockchains
    - Ethereum (mainnet, sepolia)
    - Tezos (mainnet, ghostnet)
    
    ## Features
    - Token querying with flexible filters
    - Ownership tracking across multiple chains
    - Provenance event history
    - Metadata enrichment from vendor APIs
    - Token indexing triggers
    - Metadata refresh triggers
    
    ## Authentication
    Some endpoints require authentication:
    - REST: `POST /api/v1/tokens/addresses/index` (trigger token indexing by owner addresses with job tracking)
    - GraphQL: `triggerAddressIndexing` mutation (queries, `triggerTokenIndexing`, and `triggerMetadataIndexing` mutations are public)
    
    Two authentication methods are supported:
    
    ### JWT (Bearer Token)
    Use the `Authorization: Bearer <token>` header with a valid JWT token signed with RSA (RS256/RS384/RS512).
    
    ### API Key
    Use the `Authorization: ApiKey <key>` header with a valid API key.
    
    Contact your system administrator to obtain authentication credentials.
  version: 1.0.0
  contact:
    name: Feral File
    url: https://feralfile.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/tokens/{cid}:
    get:
      summary: Get token by CID
      description: |
        Retrieves a single token by its canonical token identifier (CID).
        
        The token CID format is: `{chain}:{standard}:{contract_address}:{token_number}`
        
        Examples:
        - `eip155:1:erc721:0xabc123:1234`
        - `tezos:mainnet:fa2:KT1abc123:5678`
        
        ## Expansions
        You can expand related data using the `expand` query parameter:
        - `metadata` - Include token metadata
        - `owners` - Include paginated owner information
        - `provenance_events` - Include paginated provenance events
        - `enrichment_source` - Include vendor enrichment data
        - `display` - Include unified display field. Merges metadata and enrichment source (enrichment
          takes precedence), then applies health filtering per media-source group: for each URL field
          (`animation_url`, `image_url`), only a URL whose `token_media_health` row is `healthy` is
          returned. URLs with `broken` or `unknown` (unverified) status are omitted. If no health rows
          exist yet for a given field's source groups the raw merged value for that field is returned
          unchanged (i.e. filtering is per-field, not token-wide — an `image_url` health row does not
          affect `animation_url` pass-through and vice versa).
        - `media_asset` - Include media assets based on requested expansions
        
        **Note**: The `media_asset` expansion behavior depends on other requested expansions:
        - With `metadata` only: includes media assets from metadata
        - With `enrichment_source` only: includes media assets from enrichment source
        - With `display` only: includes media assets for the health-filtered display URLs only
        - With `metadata` + `enrichment_source`: includes media assets from both sources
        - With `metadata` + `display`: includes media from metadata + health-filtered display URLs
        - Without any source expansion: returns empty array
        
        ## Sub-resource Pagination
        When expanding owners or provenance_events, you can control pagination:
        - `owners.limit` - Limit number of owners (default: 20, max: 255)
        - `owners.offset` - Offset for owners pagination
        - `provenance_events.limit` - Limit number of events (default: 20, max: 255)
        - `provenance_events.offset` - Offset for events pagination
        - `provenance_events.order` - Sort order for events (asc/desc, default: desc)
      operationId: getToken
      tags:
        - tokens
      parameters:
        - name: cid
          in: path
          required: true
          description: Token CID in format `{chain}:{standard}:{contract_address}:{token_number}`
          schema:
            type: string
            pattern: '^(eip155:\d+|tezos:(mainnet|ghostnet)):(erc721|erc1155|fa2):.+:\d+$'
          example: "eip155:1:erc721:0xabc123:1234"
        - name: expand
          in: query
          description: Comma-separated list of expansions to include
          schema:
            type: array
            items:
              type: string
              enum: [metadata, owners, provenance_events, enrichment_source, display, media_asset]
          style: form
          explode: false
          example: ["display", "media_asset"]
        - name: owners.limit
          in: query
          description: Maximum number of owners to return (max 255)
          schema:
            type: integer
            minimum: 1
            maximum: 255
            default: 20
        - name: owners.offset
          in: query
          description: Offset for owners pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: provenance_events.limit
          in: query
          description: Maximum number of provenance events to return (max 255)
          schema:
            type: integer
            minimum: 1
            maximum: 255
            default: 20
        - name: provenance_events.offset
          in: query
          description: Offset for provenance events pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: provenance_events.order
          in: query
          description: Sort order for provenance events by timestamp
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Token found and returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad request - Invalid token CID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '422':
          description: Validation error - Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /api/v1/tokens:
    get:
      summary: List tokens
      description: |
        Retrieves a paginated list of tokens with optional filtering and sorting.
        
        ## Filters
        Multiple filter values can be provided for each filter parameter.
        All filters are applied with AND logic between different filter types,
        and OR logic within the same filter type.
        
        Examples:
        - Filter by owners: `?owner=0xabc123&owner=0xdef456`
        - Filter by chain: `?chain=eip155:1&chain=tezos:mainnet`
        - Filter by contract: `?contract_address=0xabc123`
        - Filter by token number: `?token_number=1234&token_number=5678`
        - Filter by token ID: `?token_id=12345&token_id=67890`
        - Filter by token CID: `?token_cid=eip155:1:erc721:0xabc123:1234`
        - Filter by release: `?release_id=42`
        - Combine filters: `?owner=0xabc123&chain=eip155:1`
        
        ## Sorting
        Control the sort order of results:
        - `sort_by` - Field to sort by (default: `latest_provenance`)
          - `created_at` - Sort by token creation timestamp
          - `latest_provenance` - Sort by latest provenance event timestamp (respects owner filter when provided) (default)
          - `mint_number` - Sort by authoritative mint order within a release (requires `release_id`)
        - `sort_order` - Sort direction (default: `desc`)
          - `asc` - Ascending order (oldest first)
          - `desc` - Descending order (newest first)
        
        Examples:
        - Sort by creation date, newest first: `?sort_by=created_at&sort_order=desc`
        - Sort by creation date, oldest first: `?sort_by=created_at&sort_order=asc`
        - Sort by latest activity: `?sort_by=latest_provenance&sort_order=desc` (default)
        - Sort by latest activity for specific owner: `?owner=0xabc123&sort_by=latest_provenance&sort_order=desc`
        
        ## Pagination
        Use `limit` and `offset` for pagination:
        - `limit` - Number of items per page (default: 20, max: 255)
        - `offset` - Starting position (default: 0)
        
        ## Expansions
        You can expand related data using the `expand` query parameter:
        - `metadata` - Include token metadata
        - `owners` - Include owner information (first 20 items per token)
        - `provenance_events` - Include provenance events (first 20 items per token, ordered by timestamp DESC)
        - `enrichment_source` - Include vendor enrichment data
        - `display` - Include unified display field. Merges metadata and enrichment source (enrichment
          takes precedence), then applies health filtering per media-source group: for each URL field
          (`animation_url`, `image_url`), only a URL whose `token_media_health` row is `healthy` is
          returned. URLs with `broken` or `unknown` (unverified) status are omitted. If no health rows
          exist yet for a given field's source groups the raw merged value for that field is returned
          unchanged (i.e. filtering is per-field, not token-wide — an `image_url` health row does not
          affect `animation_url` pass-through and vice versa).
        - `media_asset` - Include media assets based on requested expansions
        - `owner_provenances` - Include owner provenance history (fixed page size per token; not available on GET /tokens/{cid})
        
        **Note**: The `media_asset` expansion behavior depends on other requested expansions:
        - With `metadata` only: includes media assets from metadata
        - With `enrichment_source` only: includes media assets from enrichment source
        - With `display` only: includes media assets for the health-filtered display URLs only
        - With `metadata` + `enrichment_source`: includes media assets from both sources
        - With `metadata` + `display`: includes media from metadata + health-filtered display URLs
        - Without any source expansion: returns empty array
        
        **Important**: For list queries, `owners` and `provenance_events` return a fixed cap of items per token. Use `GET /api/v1/tokens/{cid}` for paginated owners and provenance events.
      operationId: listTokens
      tags:
        - tokens
      parameters:
        - name: owner
          in: query
          description: Filter by owner address(es) - supports multiple values
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["0xabc123", "tz1abc123"]
        - name: chain
          in: query
          description: Filter by blockchain chain(s) - supports multiple values
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Chain'
          style: form
          explode: true
          example: ["eip155:1"]
        - name: contract_address
          in: query
          description: Filter by contract address(es) - supports multiple values
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["0xabc123"]
        - name: token_id
          in: query
          description: Filter by internal database token ID(s) - supports multiple values
          schema:
            type: array
            items:
              type: integer
              format: uint64
          style: form
          explode: true
          example: [12345, 67890]
        - name: token_number
          in: query
          description: Filter by token number(s) within contract - supports multiple values
          schema:
            type: array
            items:
              type: string
              pattern: '^\d+$'
          style: form
          explode: true
          example: ["1234", "5678"]
        - name: token_cid
          in: query
          description: Filter by token CID(s) - supports multiple values
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          example: ["eip155:1:erc721:0xabc123:1234"]
        - name: release_id
          in: query
          description: Filter by internal release ID
          schema:
            type: integer
            format: int64
          example: 42
        - name: include_unviewable
          in: query
          description: Include tokens with unviewable media (no accessible media URLs). Default is false - only tokens with accessible media are returned.
          schema:
            type: boolean
            default: false
          example: false
        - name: sort_by
          in: query
          description: |
            Field to sort tokens by.
            
            Sort options:
            - `created_at` - Sort by token creation timestamp
            - `latest_provenance` - Sort by latest provenance event timestamp. When owner filter is provided, sorts by the latest provenance event for those specific owners. Without owner filter, uses the token's last provenance timestamp. (default)
            - `mint_number` - Sort by mint order within a release (requires `release_id`)
          schema:
            type: string
            enum: [created_at, latest_provenance, mint_number]
            default: latest_provenance
          example: latest_provenance
        - name: sort_order
          in: query
          description: Sort order (ascending or descending)
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          example: desc
        - name: limit
          in: query
          description: Maximum number of tokens to return (max 255)
          schema:
            type: integer
            minimum: 1
            maximum: 255
            default: 20
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: expand
          in: query
          description: |
            Comma-separated list of expansions to include.
            Note: owners and provenance_events expansions return first 20 items per token without pagination support.
            Use GET /api/v1/tokens/{cid} for paginated results.
          schema:
            type: array
            items:
              type: string
              enum: [metadata, owners, owner_provenances, provenance_events, enrichment_source, display, media_asset]
          style: form
          explode: false
          example: ["display", "media_asset"]
      responses:
        '200':
          description: Tokens retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenListResponse'
        '422':
          description: Validation error - Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
...
  /api/v1/releases/{id}:
    get:
      summary: Get release by internal id
      description: |
        Retrieves a cross-vendor release (Feral File series or Art Blocks project) and its member tokens sorted by mint number.
      operationId: getRelease
      tags:
        - releases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 42
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 255
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: expand
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [metadata, owners, owner_provenances, provenance_events, enrichment_source, display, media_asset]
          style: form
          explode: false
      responses:
        '200':
          description: Release with mint-ordered member tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseResponse'
        '404':
          description: Release not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /api/v1/tokens/index:
    post:
      summary: Trigger token indexing by CIDs
      description: |
        Triggers indexing workflows for one or more tokens by their CIDs.
        
        This endpoint accepts a list of token CIDs (up to 20) and will re-index those specific tokens,
        including metadata, provenance events, and media assets.
        
        **This endpoint does not require authentication** - it is open for public use.
        
        ## Response
        Returns a Postgres queue `job_id` plus deprecated `workflow_id` (decimal string of `job_id`) and
        `run_id` (always null) for legacy clients; use `job_id` with `GET /api/v1/jobs/{job_id}` for status.
      operationId: triggerTokenIndexing
      tags:
        - tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerTokenIndexingRequest'
      responses:
        '202':
          description: Indexing triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerIndexingResponse'
        '422':
          description: Validation error - Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using RSA signing algorithms (RS256, RS384, or RS512).
        
        Include the JWT token in the Authorization header:
        ```
        Authorization: Bearer <your-jwt-token>
        ```
        
        The token must include standard claims:
        - `exp` (expiration time) - Token must not be expired
        - `nbf` (not before) - Token must be currently valid
        - `sub` (subject) - Optional, identifies the authenticated user/service
        
        Example:
        ```
        Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API Key authentication.
        
        Include the API key in the Authorization header with 'ApiKey' prefix:
        ```
        Authorization: ApiKey <your-api-key>
        ```
        
        Example:
        ```
        Authorization: ApiKey abc123def456...
        ```